movieDetails files organized
