+++ src/rootSaga.js
@@ -2,7 +2,7 @@ import { all } from "redux-saga/effects";
 import { moviesSaga } from "./features/popularMovies/moviesSaga";
 import { peopleSaga } from "./features/popularPeople/peopleSaga";
 import { personSaga } from "./features/personDetails/personSaga";
-import { movieDetailsSaga } from "./features/movieDetailsSaga";
+import { movieDetailsSaga } from "./features/movieDetails/movieDetailsSaga";
 
 export default function* rootSaga() {
     yield all([
